chore: Move folders around to match the naming convention; Add hardhat specific formatOmniTransaction
