+++ packages/devtools-evm-hardhat/src/index.ts
@@ -7,5 +7,5 @@ export * from './internal'
 export * from './omnigraph'
 export * from './provider'
 export * from './runtime'
-export * from './signer'
+export * from './transactions'
 export * from './types'
