--- packages/devtools-evm-hardhat/src/signer/index.ts
@@ -1,1 +1,0 @@
-export * from './factory'
