+++ packages/devtools-evm-hardhat/src/transactions/format.ts
@@ -1,0 +1,9 @@
+import { formatOmniTransaction as formatOmniTransactionBase, type OmniTransaction } from '@layerzerolabs/devtools'
+import { getNetworkNameForEid } from '@/runtime'
+
+export const formatOmniTransaction = (
+    transaction: OmniTransaction
+): Record<string, string | number | bigint | undefined> => ({
+    Network: getNetworkNameForEid(transaction.point.eid),
+    ...formatOmniTransactionBase(transaction),
+})
